:fire: Remove environment from present and pop function
